+++ packages/bluesky_chat/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Note
 
+## v0.1.2
+
+- Rename from `ConvoMessage` to `ConvoMessageInput`. ([#1463](https://github.com/myConsciousness/atproto.dart/issues/1463))
+- Add `convoId` parameter on `ModerationService.getMessageContext`. ([#1463](https://github.com/myConsciousness/atproto.dart/issues/1463))
+- Add `convoId` property on `ConvoServiceRef`. ([#1463](https://github.com/myConsciousness/atproto.dart/issues/1463))
+
 ## v0.1.1
 
 - Add `.chat` property on `ProfileAssociated` object. ([#1448](https://github.com/myConsciousness/atproto.dart/issues/1448))